requirements.txt
langchain
langchain-huggingface
langchain-community
langgraph
sentence-transformers
faiss-cpu
pypdf
python-dotenv
pydantic